Bfi (bfi v0.1.0)

Brainf*ck Interpreter in Elixir.

Wanted to try my hand at parsing a language. Imperatively, Brainf*ck is trivial to interpret, which is why I wanted to try it in a functional language like Elixir.

Examples

Linux

  $> ./bfi hello.txt
  Hello world!
  $> ./bfi --eval --size=10 "++++++[>+++++++++<-]>.+++."
  69

Windows

  $> escript bfi hello.txt
  Hello world!
  $> escript bfi --eval --size=10 "++++++[>+++++++++<-]>.+++."
  69

Functions

eval(list, state)

Evaluation function that comes after the parsing step.

Uses tail-recursion with the state as the accumulator.

State

The state is a 3-tuple, composed of the data pointer dp, the memory mem, and the input queue input.

Notes

Elixir raises an error when trying to :print ASCII values that are greater than 128.

Inputting more than one character after an evaluation on :input results in queueing the rest of the input for subsequent :input calls.

When evaluating :next and :prev, note that dp is bounded by 0 and the size of mem.

When evaluating :increment and :decrement, note that negative values underflow to 255, and numbers greater than 255 overflow back to 0.

Examples

iex> Bfi.parse("++++++[>++++++++<-]>.>++++++++++.") |> Bfi.eval(Bfi.memory(3))
0
{2, {:array, 3, 0, 0, {0, 48, 10, 0, 0, 0, 0, 0, 0, 0}}, ""}
iex> Bfi.parse(",.,.>++++++++++.") |> Bfi.eval(Bfi.memory(3))
01
01
{1, {:array, 3, 0, 0, {49, 10, 0, 0, 0, 0, 0, 0, 0, 0}}, ""}

main(args)

memory(size)

Creates the initial state for eval/2, based on a given size

Examples

iex> init = Bfi.memory(3)
iex> Bfi.parse("++++++[>++++++++<-]>.") |> Bfi.eval(init)
{1, {:array, 3, 0, 0, {0, 48, 0, 0, 0, 0, 0, 0, 0, 0}}, ""}
iex> Bfi.parse("-[>-[-]]<") |> Bfi.eval(Bfi.memory(5))
{0, {:array, 5, 0, 0, {255, 0, 0, 0, 0, 0, 0, 0, 0, 0}}, ""}

parse(instruction, acc \\ [])

Parse characters into tokens and group loops.

Since Brainf*ck tokens are just characters, lexing can be condensed into this one function.

Character	Meaning

+	Increments the current cell

-	Decrements the current cell

>	Moves forward by one cell

<	Moves backward by one cell

.	Outputs the current cell

,	Takes an integer input

[	When current cell is 0, jump to after matching ]

]	Jump unconditionally to matching [

<other>	Read as a comment

Tokens

@env %{
  ?+ => :increment,
  ?- => :decrement,
  ?> => :next,
  ?< => :prev,
  ?. => :print,
  ?, => :input,
}

Brackets are parsed as nested lists.

Examples

iex> Bfi.parse("++.a")
[:increment, :increment, :print, :comment]
iex> Bfi.parse("-><")
[:decrement, :next, :prev]
iex> Bfi.parse("-[--].")
[:decrement, [:decrement, :decrement], :print]